# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Lara Movies API Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/movies": {
            "get": {
                "tags": [
                    "Movies"
                ],
                "summary": "Get a list of movies",
                "description": "Returns a list of movies",
                "operationId": "f1b3cb64ae74885cdcf404eebe09485f",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of movies per page (default is 5)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 5
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Current page (default is 1)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response: List of movies",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized: Invalid credentials",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "403": {
                        "description": "Forbidden: Insufficient permissions",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Not Found: Movies not found",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Movies"
                ],
                "summary": "Store a new movie",
                "description": "Stores a new movie in the database",
                "operationId": "99aa988b99ecf68d2dced90d100bff63",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreMovieRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Movie created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Movie created successfully"
                                        },
                                        "movie": {
                                            "$ref": "#/components/schemas/MovieResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/movies/{movie}": {
            "get": {
                "tags": [
                    "Movies"
                ],
                "summary": "Get a single movie by ID",
                "description": "Retrieves a single movie by its ID",
                "operationId": "af114386bfd942e4a5971042e6ba1996",
                "parameters": [
                    {
                        "name": "movie",
                        "in": "path",
                        "description": "ID of the movie to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Movie retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "movie": {
                                            "$ref": "#/components/schemas/MovieResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Movie not found",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Movies"
                ],
                "summary": "Update a movie by ID",
                "description": "Updates a movie by its ID",
                "operationId": "905ee2eb024274852d779a61b7773db0",
                "parameters": [
                    {
                        "name": "movie",
                        "in": "path",
                        "description": "ID of the movie to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateMovieRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Movie updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Movie updated successfully"
                                        },
                                        "movie": {
                                            "$ref": "#/components/schemas/MovieResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Movie not found",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Movies"
                ],
                "summary": "Delete a movie by ID",
                "description": "Deletes a movie by its ID",
                "operationId": "9152b84c359a7ea20dd945014c623fd1",
                "parameters": [
                    {
                        "name": "movie",
                        "in": "path",
                        "description": "ID of the movie to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Movie deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Movie deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Movie not found",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "StoreMovieRequest": {
                "title": "Store Movie Request",
                "description": "Request structure for storing a new movie",
                "required": [
                    "name",
                    "title",
                    "author"
                ],
                "properties": {
                    "name": {
                        "description": "The name of the movie",
                        "type": "string"
                    },
                    "title": {
                        "description": "The title of the movie",
                        "type": "string"
                    },
                    "author": {
                        "description": "The author of the movie",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "UpdateMovieRequest": {
                "title": "Update Movie Request",
                "description": "Request structure for storing a new movie",
                "required": [
                    "name",
                    "title",
                    "author"
                ],
                "properties": {
                    "name": {
                        "description": "The name of the movie",
                        "type": "string"
                    },
                    "title": {
                        "description": "The title of the movie",
                        "type": "string"
                    },
                    "author": {
                        "description": "The author of the movie",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "MovieResource": {
                "title": "Movie Resource",
                "description": "Movie Resource",
                "properties": {
                    "id": {
                        "description": "The unique identifier of the movie",
                        "type": "integer"
                    },
                    "name": {
                        "description": "The name of the movie",
                        "type": "string"
                    },
                    "title": {
                        "description": "The title of the movie",
                        "type": "string"
                    },
                    "author": {
                        "description": "The author of the movie",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}